Add missed includes for standard C APIs

+++ Source/Server/Server/GameService/GameManagers/Signs/SignManager.cpp
@@ -20,6 +20,8 @@
 #include "Core/Utils/File.h"
 #include "Core/Utils/Strings.h"
 
+#include <cmath>
+
 SignManager::SignManager(Server* InServerInstance, GameService* InGameServiceInstance)
     : ServerInstance(InServerInstance)
     , GameServiceInstance(InGameServiceInstance)
@@ -476,7 +478,7 @@ MessageHandleResult SignManager::Handle_RequestGetRightMatchingArea(GameClient* 
     // Normalize the values to the 0-5 range the client expects and return them.
     for (auto Pair : PotentialAreas)
     {
-        int NormalizedPopulation = (int)std::ceilf((Pair.second / (float)MaxAreaPopulation) * 5.0f);
+        int NormalizedPopulation = (int)std::ceil((Pair.second / (float)MaxAreaPopulation) * 5.0f);
 
         Frpg2RequestMessage::RequestGetRightMatchingAreaResponse_Area_info& Info = *Response.add_area_info();
         Info.set_online_area_id((uint32_t)Pair.first);
